Change WaveEquation1D example to integrationtests and moved it
