+++ applications/examples/c++/CMakeLists.txt
@@ -33,7 +33,6 @@ foreach(test
     SparseDirectSolverMUMPS
     Temperature1D
     Truss1D2N
-    WaveEquation1D
     )
 
     add_executable(${test} EXCLUDE_FROM_ALL ${test}.cpp)
